fix: provisioner and controller bug fix (#115)
* fix: provisioner and controller bug fix

* fix: make GetMaxSharesAndMaxMountReplicaCount function more defensive

+++ pkg/controller/azvolume.go
@@ -138,8 +138,13 @@ func (r *ReconcileAzVolume) triggerCreate(ctx context.Context, volumeName string
 		klog.Errorf("failed to get AzVolume (%s): %v", volumeName, err)
 		return err
 	}
-	var response *v1alpha1.AzVolumeStatusParams
 
+	//Register finalizer
+	if azVolume, err = r.initializeMeta(ctx, volumeName, azVolume, true); err != nil {
+		return err
+	}
+
+	var response *v1alpha1.AzVolumeStatusParams
 	if azVolume.Status.State == v1alpha1.VolumeOperationPending || azVolume.Status.State == v1alpha1.VolumeCreating || azVolume.Status.State == v1alpha1.VolumeCreated {
 		if azVolume.Status.State == v1alpha1.VolumeOperationPending {
 			if azVolume, err = r.updateState(ctx, volumeName, azVolume, v1alpha1.VolumeCreating); err != nil {
@@ -155,11 +160,6 @@ func (r *ReconcileAzVolume) triggerCreate(ctx context.Context, volumeName string
 				return derr
 			}
 			_, err = r.updateStatusWithError(ctx, volumeName, azVolume, err)
-			return err
-		}
-
-		//Register finalizer
-		if azVolume, err = r.initializeMeta(ctx, volumeName, azVolume, true); err != nil {
 			return err
 		}
 
@@ -183,12 +183,12 @@ func (r *ReconcileAzVolume) triggerDelete(ctx context.Context, volumeName string
 	}
 
 	// Delete all AzVolumeAttachment objects bound to the deleted AzVolume
-	volRequirement, err := labels.NewRequirement(VolumeNameLabel, selection.Equals, []string{azVolume.Spec.UnderlyingVolume})
+	volRequirement, err := labels.NewRequirement(azureutils.VolumeNameLabel, selection.Equals, []string{azVolume.Spec.UnderlyingVolume})
 	if err != nil {
 		return err
 	}
 	if volRequirement == nil {
-		return status.Error(codes.Internal, fmt.Sprintf("Unable to create Requirement to for label key : (%s) and label value: (%s)", VolumeNameLabel, azVolume.Spec.UnderlyingVolume))
+		return status.Error(codes.Internal, fmt.Sprintf("Unable to create Requirement to for label key : (%s) and label value: (%s)", azureutils.VolumeNameLabel, azVolume.Spec.UnderlyingVolume))
 	}
 
 	labelSelector := labels.NewSelector().Add(*volRequirement)
@@ -201,18 +201,25 @@ func (r *ReconcileAzVolume) triggerDelete(ctx context.Context, volumeName string
 
 	klog.V(5).Infof("number of attachments found: %d", len(attachments.Items))
 	for _, attachment := range attachments.Items {
-		if err = r.azVolumeClient.DiskV1alpha1().AzVolumeAttachments(r.namespace).Delete(ctx, attachment.Name, metav1.DeleteOptions{}); err != nil && !errors.IsNotFound(err) {
-			klog.Errorf("failed to delete AzVolumeAttachment (%s): %v", attachment.Name, err)
-			return err
+		if attachment.DeletionTimestamp == nil {
+			if err = r.azVolumeClient.DiskV1alpha1().AzVolumeAttachments(r.namespace).Delete(ctx, attachment.Name, metav1.DeleteOptions{}); err != nil && !errors.IsNotFound(err) {
+				klog.Errorf("failed to delete AzVolumeAttachment (%s): %v", attachment.Name, err)
+				return err
+			}
+			klog.V(5).Infof("Set deletion timestamp for AzVolumeAttachment (%s)", attachment.Name)
 		}
-		klog.V(5).Infof("Set deletion timestamp for AzVolumeAttachment (%s)", attachment.Name)
+	}
+
+	if len(attachments.Items) > 0 {
+		return status.Errorf(codes.Aborted, "volume deletion requeued until attached azVolumeAttachments are entirely detached...")
 	}
 
 	// only try deleting underlying volume 1) if volume creation was successful and 2) volumeDeleteRequestAnnotation is present
 	// if the annotation is not present, only delete the CRI and not the underlying volume
 	if azVolume.Annotations != nil {
 		if _, ok := azVolume.Annotations[azureutils.VolumeDeleteRequestAnnotation]; ok &&
-			(azVolume.Status.State == v1alpha1.VolumeCreated || azVolume.Status.State == v1alpha1.VolumeCreationFailed || azVolume.Status.State == v1alpha1.VolumeDeleting) {
+			azVolume.Status.State == v1alpha1.VolumeCreating || azVolume.Status.State == v1alpha1.VolumeCreated || azVolume.Status.State == v1alpha1.VolumeDeleting || azVolume.Status.State == v1alpha1.VolumeDeletionFailed {
+
 			if azVolume, err = r.updateState(ctx, volumeName, azVolume, v1alpha1.VolumeDeleting); err != nil {
 				return err
 			}
@@ -237,7 +244,7 @@ func (r *ReconcileAzVolume) triggerDelete(ctx context.Context, volumeName string
 	}
 
 	// Update status of the object
-	if _, err = r.updateStatus(ctx, azVolume.Name, azVolume, azVolume.Status.Detail.Phase, true, nil); err != nil {
+	if _, err = r.updateStatus(ctx, azVolume.Name, azVolume, v1alpha1.VolumeReleased, true, nil); err != nil {
 		return err
 	}
 
@@ -355,6 +362,9 @@ func (r *ReconcileAzVolume) expandVolume(ctx context.Context, azVolume *v1alpha1
 }
 
 func (r *ReconcileAzVolume) createVolume(ctx context.Context, azVolume *v1alpha1.AzVolume) (*v1alpha1.AzVolumeStatusParams, error) {
+	if azVolume.Status.Detail != nil && azVolume.Status.Detail.ResponseObject != nil {
+		return azVolume.Status.Detail.ResponseObject, nil
+	}
 	// use deep-copied version of the azVolume CRI to prevent any unwanted update to the object
 	copied := azVolume.DeepCopy()
 	return r.cloudProvisioner.CreateVolume(ctx, copied.Spec.UnderlyingVolume, copied.Spec.CapacityRange, copied.Spec.VolumeCapability, copied.Spec.Parameters, copied.Spec.Secrets, copied.Spec.ContentVolumeSource, copied.Spec.AccessibilityRequirements)
